Remove some unnecessary casts

+++ src/net/ssehub/kernel_haven/code_model/CodeModelProvider.java
@@ -21,7 +21,7 @@ public class CodeModelProvider extends AbstractProvider<SourceFile> {
 
     @Override
     protected long getTimeout() {
-        return (long) config.getValue(DefaultSettings.CODE_PROVIDER_TIMEOUT);
+        return config.getValue(DefaultSettings.CODE_PROVIDER_TIMEOUT);
     }
     
     @Override
